requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# ML / Data
xgboost==2.0.3
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4

# Explainability
shap==0.45.1

# Serialization
joblib==1.4.2

# Gemini (Google Generative AI)
google-generativeai==0.7.2

# Validation & typing
pydantic==2.7.1

# Optional (useful for dev/debugging)
matplotlib==3.9.0